Fixed some bugs.
Fixed some warnings in PaletteEditor.

Fixed the relation code to use user_data specific to the table the relation is for. Using a generic user_data for all relation lists was causing a relation_updated to trigger a refresh() on all relation lists (some of which did not have a table set).

Fixed a message dialog title in the skill list.